Bugfix/testing: The unit-test for ClearBaseFilter failed when the folder for the database did not exist before run

+++ trunk/Storage/test/dk/statsbiblioteket/summa/storage/api/filter/ClearBaseFilterTest.java
@@ -72,7 +72,9 @@ public class ClearBaseFilterTest extends TestCase {
     public Storage createTestStorage() throws Exception {
         String dbLocation = "summatest" + File.separator + "testDB";
         File dbFile = new File(dbLocation);
-        Files.delete(dbFile.getParentFile());
+        if (dbFile.getParentFile().exists()) {
+            Files.delete(dbFile.getParentFile());
+        }
 
         storage = StorageFactory.createStorage(
                         Configuration.newMemoryBased(
